remove dependencies of /src from headers
